+++ src/da4ml/_binary/cmvm/api.hh
@@ -4,10 +4,10 @@
 #include <string>
 #include <vector>
 
-double minimal_latency(
+float minimal_latency(
     const xt::xarray<float> &kernel,
     const std::vector<QInterval> &qintervals,
-    const std::vector<double> &latencies,
+    const std::vector<float> &latencies,
     int carry_size = -1,
     int adder_size = -1
 );
@@ -19,7 +19,7 @@ PipelineResult _solve(
     int hard_dc = -1,
     int decompose_dc = -2,
     const std::vector<QInterval> &qintervals = {},
-    const std::vector<double> &latencies = {},
+    const std::vector<float> &latencies = {},
     int adder_size = -1,
     int carry_size = -1
 );
@@ -31,7 +31,7 @@ PipelineResult solve(
     int hard_dc = -1,
     int decompose_dc = -2,
     const std::vector<QInterval> &qintervals = {},
-    const std::vector<double> &latencies = {},
+    const std::vector<float> &latencies = {},
     int adder_size = -1,
     int carry_size = -1,
     bool search_all_decompose_dc = true
